Refactor download, merge in one file to have better import in the package
