+++ desed/desed/__init__.py
@@ -1,5 +1,4 @@
-from .download_real import download_real
-from .download_soundbank import download_soundbank
+from desed.download import download_real, download_soundbank
 from .generate_synthetic import (
     SoundscapesGenerator,
     generate_files_from_jams,
