+++ examples/generate_dcase_task4_2020.py
@@ -4,8 +4,9 @@ import os
 from pprint import pformat
 
 import desed
+import desed.download
 from desed.utils import download_and_unpack_archive
-from desed.download_soundbank import split_soundbank_train_val
+from desed.download import split_desed_soundbank_train_val
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -24,18 +25,18 @@ if __name__ == "__main__":
     # ##########
     # Real data
     # ##########
-    desed.download_real(real_basedir)
+    desed.download.download_real(real_basedir)
 
     # ##########
     # Synthetic soundbank
     # ##########
     # Download the soundbank if needed
     if not os.path.exists(soundbank_dir):
-        desed.download_soundbank(soundbank_dir, sins_bg=True, tut_bg=True)
+        desed.download.download_soundbank(soundbank_dir, sins_bg=True, tut_bg=True)
     else:
         # If you don't have the validation split, rearrange the soundbank in train-valid (split in 90%/10%)
         if not os.path.exists(os.path.join(soundbank_dir, "audio", "validation")):
-            split_soundbank_train_val(soundbank_dir)
+            split_desed_soundbank_train_val(soundbank_dir)
 
     # ##########
     # Synthetic soundscapes
